deepseek-ai-web-crawler/requirements.txt
beautifulsoup4>=4.12.2
python-dotenv==1.0.1
python-slugify==8.0.1
requests==2.32.3
selenium==4.15.2
undetected-chromedriver==3.5.4
pandas>=2.0.0,<2.1.0
pydantic==2.9.2
crawl4ai==0.3.0
asyncio==3.4.3
aiohttp>=3.9.1
tqdm==4.66.5
playwright>=1.30.0
Pillow>=9.0.0
openpyxl>=3.1.2